python/packages/sdk/pyproject.toml
[project]
name = "zpools-sdk"
version = "0.1.0"
description = "Python SDK for zpools.io"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx",
    "pydantic",
    "attrs>=21.3.0",
    "python-dateutil",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zpools"]
